fix typed_dictionary compile-time regression
